requirements.txt
Django==5.0.2
django-extensions==3.2.3
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
django-tinymce==3.7.1
pillow==10.2.0
django-storages==1.14.2
django-cors-headers==4.3.1
pygraphviz==1.12
psycopg2==2.9.9 #postgresql dependent
whitenoise==6.6.0
dj-database-url==2.1.0
python-dotenv==1.0.1
requests==2.31.0
# django-bootstrap5==23.4
# boto3==1.34.47